Frontend/requirements.txt
streamlit
requests
pandas
streamlit
requests
pandas
PyPDF2
python-docx
